backend/requirements.txt
Django>=3.0,<4.0
djangorestframework >=3.10
markdown
django-filter
drf-yasg
psycopg2-binary>=2.8